(no verify) fix missing embed (whoops), working on merged table types
