+++ pkg/storage/system_intake.go
@@ -576,6 +576,7 @@ func (s *Store) GetSystemIntakesWithLCIDs(ctx context.Context) ([]*models.System
 
 func (s *Store) GetMySystemIntakes(ctx context.Context) ([]*models.SystemIntake, error) {
 	var intakes []*models.SystemIntake
+
 	err := namedSelect(ctx, s, &intakes, sqlqueries.SystemIntake.GetByUser, args{
 		"eua_user_id": appcontext.Principal(ctx).Account().Username,
 	})
